+++ CHANGES.md
@@ -1,5 +1,7 @@
 0.8.0
 -----
+* Futures no longer use a thread unless created with a block
+* No more future thread-leaks! Future threads auto-terminate now
 * Rename Celluloid#async to Celluloid#defer
 * Celluloid#tasks now returns an array of tasks with a #status attribute
 * Reduce coupling between Celluloid and DCell. Breaks compatibility with